refactor(GuideDetailSection): moved parser logic
